perf: ffi: don't clone ClientId
Instead of cloning an inner value of an Arc we should use .as_ref() to
retrieve it.

For ClientId we then implement AsRef<core_crypto::ClientIdRef> which
allows for a zero-copy conversion of core_crypto_ffi::Clientd to
core_crypto::ClientIdRef and thus to core_crypto::ClientId.

We also change the function signature of get_device_ids() to accept a
slice of ClientIdRefs.