feat(keystore): add the ability to borrow a primary key to `trait BorrowedPrimaryKey`
This takes us full circle: an entity which has a borrowed form
can retrieve itself from that borrowed form, and also can produce
that borrowed form.

I went back and forth several times about whether this should be its
own distinct trait. I have an intuition that there will be
occasional cases where we need one part of this, but the other part
is undesirable or impossible. However, I couldn't demonstrate any
concrete case like that even after trying for a while, so I decided
to merge the traits on the basis that fewer is better. If we do encounter
a concrete case where they need to be split, we can handle that then.