refactor: decouple session creation from client construction
The idea here is to create an "empty" CoreCrypto client first that holds
no session and no pki environment. We create the mls Session on mls init
which allows us to get rid of the inner session and provide transport as
part of mls_init.
The keystore should live in CoreCrypto since it is shared between
proteus and mls. Proteus should not need to know anything about the mls
session.