refactor: adjust transaction context to new mls session
proteus transactions should not rely on the the mls crypto provider to
access the keystore.

mls members are only available when the mls session exists, so lets
access them through the mls session.

I think there is a design flaw in the transaction context that we should
address later.
The transaction context never knows if it is a mls or proteus
transaction. There should probably be separate transaction contexts so
that a transaction context just uses its session instead of checking for
it's existence. It also doesn't need to know anything about the other
session.

However, for now we keep it like that.

Also setting transport on an already initialized session is no longer
possible. It only existed for tests which we will have to adjust.