chore(crypto): make `CertificatePrivateKey.signature_scheme` private
Rustc 1.92 seems to think that we never read that value. Now, a human
would look at credential/x509.rs:81 and say "that sure looks like the
signature scheme is being read!" But not rustc 1.92. (This is probably
a rust bug.)

So this commit tightens up `CertificatePrivateKey`: it now has a
constructor, only private fields, and a getter for the signature
scheme. And in fact this does enable the compiler to notice that
we read the signature scheme at some point after creating it.