feat(keystore): introduce `trait OwnedKeyType`
We already knew that there is a real difference between key types
which are owned and those which are borrowed. We already knew that
entity keys need to be of the owned kind.

Adding this trait allows us to formalize that notion, and add
a utility to construct an owned variant from some arbitrary bytes.