refactor(keystore): hold onto entity items in dynamic dispatch
We used to do some crazy thing involving unencrypted serialization
for some reason. It's much simpler this way: when an entity turns
itself into a dynamic entity, we just keep it. We do Arc-wrap it,
because there are times we want to iterate over a bunch of owned
instances and don't want the expense of cloning these potentially-
large items every time. But from there the API is straightforward.

While it was cool we could make the borrowing approach work for
dynamic entities, in practice it turned out that in general when
we hand an entity to the database to save, we really want it to
take over ownership, which means this pattern is better.