chore(keystore): impl `BorrowPrimaryKey` for `MlsPendingMessage`
Pending messages have no real primary key and are always/only accessed
and removed in bulk, which turns out to be _really annoying_ with regard
to the new entity traits, which assumes that some primary key does exist.

For now we're faking the foreign id as the primary key so that encryption
works at all, and also for dynamic dispatch. But I'd really love to
rework this at some point in the future.