refactor(keystore): simplify parameter type for borrowing trait methods
The traits didn't previously allow for multiple distinct implementations
of borrowing types sharing an owned type (I think) anyway, and we
already had an associated type naming a specific borrowed primary
key, so it was pointless overkill to accept a complicated generic
parameter type in these methods. Better and simpler to just use the
associated type.