refactor(keystore): break out primary key from entity type
So, I thought there were no more trait refactors incoming. But then
it turned out that there was a bug in the auto-implementations
of the unique entity helper on wasm only: `Entity` had an auto-impl
where `T: UniqueEntity` (so it could reference `T::KEY`), but
`UniqueEntity` had an auto-impl where `T: Entity` (so it could reference
`T::PrimaryKey`), and rustc was running out of stack trying to figure
out if a particular instance of `T` implemented both or neither of those.

So we broke out `PrimaryKey` from `Entity` which broke that loop,
and as a nice side effect means that there's now better symmetry
between `PrimaryKey` and `BorrowPrimaryKey`, and `Entity` and `EntityGetBorrowed`
(as well as `EntityDatabaseMutation` and `EntityDeleteBorrowed`).

But now I have to fix all the impls.