chore(keystore): use correct primary key for mls types
For some reason we're computing and storing a sha256 hash for these
types, but we're not using it as the primary key. OpenMLS uses the
raw id field as the primary key in a way that we can't modify,
so that's what has to be the pk. As for the hash: wasted space.