chore(keystore): handle `MlsPendingMessage`s properly
This entity type is really annoying because it just doesn't
have a single primary key per entity. So when we faked it as just
the conversation id, as soon as there was more than 1 pending message
per conversation, they started overwriting each other within the transaction.

So we've added a bunch of changes and workarounds to make it go.
I'm hopeful that in the future we can get rid of all or nearly all
of this, but for now, we're stuck with it.