chore(keystore): handle keys which are not byte vectors
For whatever reason the entity new derive macro was calling
`new_get` and `new_delete` with `String` or `&str` instances instead
of the expected `Vec<u8>` / `&[u8]`, and of course that was failing
type checking. But actually we'd prefer to have those functions available.
So broadening the acceptable types solves the problem for us (probably).